# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Sync API
  description: Couple's shared chat MVP with AI reflection and clarification
  version: 0.1.0
  contact:
    name: Sync Team
    email: support@sync.app
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api-staging.sync.app
    description: Staging server
  - url: https://api.sync.app
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /auth/request-code:
    post:
      tags:
        - Authentication
      summary: Request verification code
      description: Send a verification code to the user's email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequestCode'
      responses:
        '204':
          description: Verification code sent successfully
        '400':
          description: Invalid email format
        '429':
          description: Rate limit exceeded
  /auth/verify-code:
    post:
      tags:
        - Authentication
      summary: Verify code and get access token
      description: Verify the code and receive an access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthVerifyCode'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid code or email
        '401':
          description: Invalid or expired code
  /couples:
    post:
      tags:
        - Couples
      summary: Create a new couple
      description: Create a new couple relationship
      responses:
        '201':
          description: Couple created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCoupleResponse'
        '401':
          description: Unauthorized
  /invites:
    post:
      tags:
        - Invites
      summary: Create an invite
      description: Create an invite for the partner to join the couple
      responses:
        '201':
          description: Invite created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateInviteResponse'
        '401':
          description: Unauthorized
  /invites/{code}/accept:
    post:
      tags:
        - Invites
      summary: Accept an invite
      description: Accept an invite to join a couple
      security: []
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Invite accepted successfully
        '400':
          description: Invalid or expired invite code
        '404':
          description: Invite not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthRequestCode:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
    AuthVerifyCode:
      type: object
      required:
        - email
        - code
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        code:
          type: string
          minLength: 6
          maxLength: 6
          example: "123456"
    AuthResponse:
      type: object
      required:
        - accessToken
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    CreateCoupleResponse:
      type: object
      required:
        - coupleId
      properties:
        coupleId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
    CreateInviteResponse:
      type: object
      required:
        - code
        - link
      properties:
        code:
          type: string
          example: "ABC123"
        link:
          type: string
          format: uri
          example: "https://sync.app/join/ABC123"
...